refactor: asanaCard into asanaDetail
